Minimal cleanup, more serial prints for debugging purposes, changed JSON parser to dynamic memory allocation

+++ pj/flow.h
@@ -46,7 +46,7 @@ typedef struct movieframe
     char r;
     char g;
     char b;
-    short position;
+    int position;
 } movieframe;
 
 // state structure for state machine
